requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
flask_bcrypt
google-cloud-bigquery
google-auth
pandas
python-dotenv